+++ quickcheck-plutus-instances/src/Test/QuickCheck/Plutus/Instances.hs
@@ -51,7 +51,10 @@ import Plutus.V1.Ledger.Scripts (
   DatumHash (DatumHash),
   ValidatorHash (ValidatorHash),
  )
-import Plutus.V1.Ledger.Time (POSIXTime (POSIXTime))
+import Plutus.V1.Ledger.Time (
+  DiffMilliSeconds (DiffMilliSeconds),
+  POSIXTime (POSIXTime),
+ )
 import Plutus.V1.Ledger.Tx (
   TxOut (TxOut),
   TxOutRef (TxOutRef),
@@ -62,6 +65,8 @@ import Plutus.V1.Ledger.Value (
   CurrencySymbol (CurrencySymbol),
   TokenName (TokenName),
   Value (Value),
+  flattenValue,
+  singleton,
  )
 import PlutusTx (
   Data (B, Constr, I, List, Map),
@@ -76,6 +81,7 @@ import PlutusTx.Builtins.Internal (
   BuiltinString (BuiltinString),
  )
 import PlutusTx.Prelude qualified as PTx
+import PlutusTx.Ratio qualified as Ratio
 import Test.QuickCheck.Arbitrary (
   Arbitrary (arbitrary, shrink),
   Arbitrary1 (liftArbitrary, liftShrink),
@@ -90,13 +96,14 @@ import Test.QuickCheck.Gen (
   oneof,
   scale,
   sized,
+  suchThat,
   variant,
   vectorOf,
  )
 import Test.QuickCheck.Instances.Text ()
 import Test.QuickCheck.Modifiers (NonNegative (NonNegative))
 import Test.QuickCheck.Plutus.Modifiers (
-  UniqueKeys (UniqueKeys, unUniqueKeys),
+  UniqueKeys (UniqueKeys),
   UniqueList (UniqueList),
   uniqueListOf,
  )
@@ -181,6 +188,19 @@ instance Function Data where
         Right (Left ds) -> List ds
         Right (Right (Left keyVals)) -> Map keyVals
         Right (Right (Right (ix, ds))) -> Constr ix ds
+
+-- | @since 1.5
+deriving via Integer instance Arbitrary DiffMilliSeconds
+
+-- | @since 1.5
+deriving via Integer instance CoArbitrary DiffMilliSeconds
+
+-- | @since 1.5
+instance Function DiffMilliSeconds where
+  function = functionMap into DiffMilliSeconds
+    where
+      into :: DiffMilliSeconds -> Integer
+      into (DiffMilliSeconds i) = i
 
 -- | @since 1.0
 deriving via BuiltinByteString instance Arbitrary LedgerBytes
@@ -463,16 +483,10 @@ instance Arbitrary Value where
     UniqueList css <- uniqueListOf num
     lst <- forM css $ \cs -> do
       UniqueList tns <- uniqueListOf num
-      lst' <- forM tns $ \tn -> (tn,) <$> arbitrary
+      lst' <- forM tns $ \tn -> (tn,) <$> arbitrary `suchThat` (PTx./= PTx.zero)
       pure (cs, AssocMap.fromList lst')
     pure . Value . AssocMap.fromList $ lst
-  shrink (Value mp) =
-    fmap (Value . unUniqueKeys) . liftShrink shrinkAsUniqueKeys . UniqueKeys $ mp
-    where
-      shrinkAsUniqueKeys ::
-        AssocMap.Map TokenName Integer ->
-        [AssocMap.Map TokenName Integer]
-      shrinkAsUniqueKeys = fmap unUniqueKeys . shrink . UniqueKeys
+  shrink = map (foldMap (\(cs, tn, i) -> singleton cs tn i)) . shrink . flattenValue
 
 -- | @since 1.1
 deriving via
@@ -646,6 +660,22 @@ instance Function BuiltinString where
     where
       into :: BuiltinString -> Text
       into (BuiltinString t) = t
+
+-- | @since 1.4
+instance Arbitrary Ratio.Rational where
+  arbitrary = Ratio.fromGHC <$> arbitrary
+  shrink = fmap Ratio.fromGHC . shrink . Ratio.toGHC
+
+-- | @since 1.4
+instance CoArbitrary Ratio.Rational where
+  coarbitrary r gen = do
+    let num = Ratio.numerator r
+    let den = Ratio.denominator r
+    coarbitrary num . coarbitrary den $ gen
+
+-- | @since 1.4
+instance Function Ratio.Rational where
+  function = functionMap Ratio.toGHC Ratio.fromGHC
 
 -- Helpers