+++ quickcheck-plutus-instances/CHANGELOG.md
@@ -4,6 +4,25 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
 ## Unreleased
 
+## 1.5 -- 2021-12-14
+
+### Changed
+
+* Instance `Arbitrary` for `Value`:
+  prevented from generating currency with _0_ quantity
+
+### Added
+
+* Instances of `Arbitrary`, `CoArbitrary` and `Function` for
+  `DiffMilliSeconds`
+
+## 1.4 -- 2021-12-10
+
+### Added
+
+* Instances of `Arbitrary`, `CoArbitrary` and `Function` for
+  `PlutusTx.Ratio.Rational`
+
 ## 1.3 -- 2021-12-09
 
 ### Added
@@ -12,7 +31,7 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
 ### Changed
 
-* Instance Arbitrary for `Value` limited by number of AssetClasses.
+* Instance `Arbitrary` for `Value` limited by number of AssetClasses.
   Now the number equals to `(log2 testSize)^2`
 
 ## 1.2 -- 2021-11-23